refactor(conf) make separate module for conf_loader listeners
Also remove some code duplication around the parsing of listeners.
